+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Versions before 2.1.2 have been backfilled; I gotta be more on top of changelogs
 	* Not everything has been moved into the API package yet for ABI reasons.
 * Code cleanups that hopefully didn't break ABI compat
 * Remove some unused stuff from the jar
+* Vertical slab placement no longer completely sucks
 * Fix a bug where templates that look like blocks with randomized models, such as stone, could reroll their blockstate on every resource load.
   * Forgot to specify a random seed.
   * All templated blocks still use the *same* model, so templated stone will still not be randomly rotated/flipped, but at least it's now the *same* same model.